Move db handler to specific dir
